Ajustes nos tipos de dados e diretorias

+++ trabalho-pratico/src/entidades/artist.c
@@ -147,7 +147,7 @@ void freeArtist(Artist* artist){
 }
 
 // getters de artista.
-int* getArtistId(Artist* a){
+long int* getArtistId(Artist* a){
     return a->id ? &a->id : NULL;
 }
